+++ .env.example
@@ -2,5 +2,9 @@
 PG_CONTAINER=mycontainer
 PG_USER=mysuperuser
 PG_DB=mydb
-PG_PWD=mypwd
+# example password creation into sha256:
+# printf '%s' 'mypwd' | sha256sum
+# echo -n 'mypwd' | sha256sum
+# echo -n mypwd | openssl dgst -binary -sha256 | openssl base64
+PG_PWD=047c88291e9877788365c59346e6918d54529cdceb7daf1e5cd490493a8e2028 *-
 PG_TABLE=mytable
