+++ README.md
@@ -1,11 +1,17 @@
-# What does this repo do
+# What does this repo do?
 This example shows how to dockerise a postgres container.
-- I want to create my container with defined variables in a .env file.
-- I want to be able to seed my container with a pre-created table init.mytable.csv BUT NOT overwrite the existing data inside: "/var/postgresql/data/". (I want to merge it with the existing data)
+- I want to create my container with defined variables in a *.env file*.
+- I want to be able to seed my container with a pre-created table **init.mytable.csv** 
+  - I want to provide that seed - **init.mytable.csv** - with variables from *.env*
+  - I want to merge **init.mytable.csv** with the existing data inside "/var/postgresql/data/". **(NOT: overwrite the data folder!)**
+
 - I want to provide a conf file to the container onload.
-(Having to symlink to another folder to prevent timing issue of docker loading its own default postgresql.conf over my loaded item)
+(Solution used: Symlink to another folder to prevent timing issue of docker loading its own default postgresql.conf over my loaded item)
 - I want to be able to connect to the container via docker exec command and run psql statements
 - I want to be able to connect to the container exernally via nodejs or equivalent and run sql statements against it to output data
+- I want to restrict access to the docker-container to my local ip address
+- I want to restrict authentication to use sha256 to the docker-container
+- I want to create tests against the above for all expected behaviour
 ---
 # Setup
 ## 1. First copy .env.example to both folders as .env
