+++ dockerise-postgres/postgresql_data/conf/pg_hba.conf
@@ -85,16 +85,22 @@
 
 # TYPE  DATABASE        USER            ADDRESS                 METHOD
 
-# "local" is for Unix domain socket connections only
-local   all             all                                     trust
-# IPv4 local connections:
-host    all             all             127.0.0.1/32            trust
-# IPv6 local connections:
-host    all             all             ::1/128                 trust
-# Allow replication connections from localhost, by a user with the
-# replication privilege.
-local   replication     all                                     trust
-host    replication     all             127.0.0.1/32            trust
-host    replication     all             ::1/128                 trust
+# DEFUULTS
+# # "local" is for Unix domain socket connections only
+# local   all             all                                     trust
+# # IPv4 local connections:
+# host    all             all             127.0.0.1/32            trust
+# # IPv6 local connections:
+# host    all             all             ::1/128                 trust
+# # Allow replication connections from localhost, by a user with the
+# # replication privilege.
+# local   replication     all                                     trust
+# host    replication     all             127.0.0.1/32            trust
+# host    replication     all             ::1/128                 trust
 
-host all all all md5
+# Access via my computer to host (ie nodejs)
+host    mydb             mysuperuser      192.168.0.0/16         scram-sha-256
+
+# Access via my computer to docker exec -it mycontainer psql -U mysuperuser -d mydb
+
+local   all             mysuperuser               scram-sha-256
