+++ dockerise-postgres/postgresql_data/conf/postgresql.conf
@@ -58,7 +58,7 @@
 
 # - Connection Settings -
 
-listen_addresses = '*'
+listen_addresses = '*' # 192.168.1.101,localhost'
                                         # comma-separated list of addresses;
                                         # defaults to 'localhost'; use '*' for all      
                                         # (change requires restart)
